Extract Scoreboard to separate file
